# lib/api-zod/src/generated/types/blockDateRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Susruta Hospital API
 * OpenAPI spec version: 0.1.0
 */

export interface BlockDateRequest {
  date: string;
  reason?: string | null;
}