# lib/api-zod/src/generated/types/monthAvailability.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Susruta Hospital API
 * OpenAPI spec version: 0.1.0
 */

export interface MonthAvailability {
  month: string;
  isOpen: boolean;
  availableDates: string[];
  blockedDates: string[];
}