# lib/api-zod/src/generated/types/openMonth.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Susruta Hospital API
 * OpenAPI spec version: 0.1.0
 */

export interface OpenMonth {
  id: number;
  /** YYYY-MM format */
  month: string;
  isOpen: boolean;
  createdAt: string;
}