# lib/api-zod/src/generated/types/testimonial.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Susruta Hospital API
 * OpenAPI spec version: 0.1.0
 */

export interface Testimonial {
  id: number;
  patientName: string;
  patientLocation?: string | null;
  content: string;
  contentTe?: string | null;
  rating: number;
  isPublished: boolean;
  createdAt: string;
}