# lib/api-zod/src/generated/types/updateSettingsRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Susruta Hospital API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSettingsRequest {
  testimonialsEnabled?: boolean;
  appointmentBookingEnabled?: boolean;
  clinicPhone1?: string;
  clinicPhone2?: string | null;
  clinicEmail?: string | null;
  clinicAddress?: string;
  workingHours?: string | null;
}